Change depth typing

+++ src/node.hpp
@@ -10,9 +10,9 @@ template <typename T>
 struct node {
     vector<vector<T> > tiles;
     int cost;
-    int8_t depth;
+    int depth;
 
-    node(const vector<vector<T> > &tile = {{0}}, int cost = 0, int8_t depth = 0) : tiles(tile), cost(cost), depth(depth) {};
+    node(const vector<vector<T> > &tile = {{0}}, int cost = 0, int depth = 0) : tiles(tile), cost(cost), depth(depth) {};
 
     // Comparison operators for compatiblity with priority_queue
     bool operator<(const node &rhs) const {
